pyproject.toml
[project]
name = "kirocli-chatbot-gateway"
version = "0.1.0"
description = "Multi-platform chatbot gateway for Kiro CLI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "lark-oapi>=1.4.2",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
discord = [
    "discord.py>=2.3.0",
]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]
